+++ flap/src/main/java/me/yifeiyuan/flap/FlapItemPool.java
@@ -5,12 +5,13 @@ import android.support.v7.widget.RecyclerView;
 /**
  * A global RecycledViewPool that can be shared among RecyclerViews , which is enabled by default.
  *
- * @author 程序亦非猿 [Follow me](<a> https://github.com/AlanCheen</a>)
- * @version 1.0
  * @see FlapAdapter#setUseFlapItemPool(boolean)
  *
  * Flap Github: <a>https://github.com/AlanCheen/Flap</a>
+ *
+ * @author 程序亦非猿 [Follow me](<a> https://github.com/AlanCheen</a>)
  * @since 2019/1/2
+ * @since 0.9.1
  */
 public class FlapItemPool extends RecyclerView.RecycledViewPool {