+++ flap-annotations/src/main/java/me/yifeiyuan/flap/annotations/Component.java
@@ -7,7 +7,7 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.SOURCE)
 @Target(ElementType.TYPE)
-public @interface Flap {
+public @interface Component {
 
     /**
      * @return the layout res id